requirement.txt
asgiref==3.2.5
Django==2.2.1
django-bootstrap-modal-forms==1.5.0
django-summernote==0.8.11.6
django-widget-tweaks==1.4.8
Pillow==7.0.0
pkg-resources==0.0.0
pytz==2019.3
sqlparse==0.3.1
